+++ test/e2e/gmrs/anid_test.clj
@@ -1,0 +1,2 @@
+(ns gmrs.anid-test
+  (:require [gmrs.command :refer :all]))